Improved class Sphere.

+++ source/Scene.cpp
@@ -2,7 +2,7 @@
 #include "Vec3.h"
 
 void Scene::addSurface(ISurface* surface) {
-	for(int i = 0; i < nodes.size(); i++) {
+	for(size_t i = 0; i < nodes.size(); i++) {
 		if(nodes[i] == surface) {
 			return;
 		}
@@ -34,7 +34,7 @@ ISurface* Scene::getSurface(const int index) const {
 int Scene::getIntersection(const Ray & ray) const {
 	float t1, t = inf;
 	int index;
-	for(int i = 0; i < nodes.size(); i++) {
+	for(size_t i = 0; i < nodes.size(); i++) {
 		if(nodes[i]->getIntersection(ray, t1)) {
 			if(t1 < t) {
 				t = t1;
